rename masterservice to rootcoord
