+++ internal/rootcoord/root_coord_test.go
@@ -9,7 +9,7 @@
 // is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 // or implied. See the License for the specific language governing permissions and limitations under the License.
 
-package masterservice
+package rootcoord
 
 import (
 	"context"
